work on time bug
