+++ app/_components/ui/(form)/(controls)/ExerciseSelect.tsx
@@ -27,7 +27,7 @@ import { Info } from "lucide-react";
 import { useFormContext } from "react-hook-form";
 import { Separator } from "../../(shadcn)/separator";
 import ExerciseVideos from "../ExerciseVideos";
-import { Exercise } from "./ExerciseFields";
+import { Exercise } from "@/app/content";
 
 const ExerciseSelect = ({ exercise }: { exercise: Exercise }) => {
   const {
